+++ spec/spec_helper.rb
@@ -15,7 +15,7 @@ ActiveRecord::Schema.define do
   end
 end
 
-class User < ActiveRecord::Base # rubocop:disable Rails/ApplicationRecord
+class User < ActiveRecord::Base
   validates :name, presence: true
 
   def self.localized_attr_name_for(name, locale)